Working

+++ plugins/pi4j-plugin-pigpio/src/main/java/com/pi4j/plugin/pigpio/PiGpioPlugin.java
@@ -37,6 +37,7 @@ import com.pi4j.plugin.pigpio.provider.gpio.digital.PiGpioDigitalInputProvider;
 import com.pi4j.plugin.pigpio.provider.gpio.digital.PiGpioDigitalOutputProvider;
 import com.pi4j.plugin.pigpio.provider.i2c.PiGpioI2CProvider;
 import com.pi4j.plugin.pigpio.provider.pwm.PiGpioPwmProvider;
+import com.pi4j.plugin.pigpio.provider.serial.PiGpioSerialProvider;
 import com.pi4j.provider.Provider;
 
 import java.io.IOException;
@@ -101,6 +102,7 @@ public class PiGpioPlugin implements Plugin {
                 PiGpioDigitalOutputProvider.newInstance(piGpio),
                 PiGpioPwmProvider.newInstance(piGpio),
                 PiGpioI2CProvider.newInstance(piGpio),
+                PiGpioSerialProvider.newInstance(piGpio)
         };
 
         // register all PiGpio I/O Providers with the plugin service
@@ -110,6 +112,6 @@ public class PiGpioPlugin implements Plugin {
     @Override
     public void shutdown(Context context) throws IOException {
         // shutdown the PiGpio library
-        piGpio.terminate();
+        if(piGpio != null) piGpio.terminate();
     }
 }
